fix example tests

+++ examples/databricks/transactions/silver/schemas/clean_transactions.py
@@ -1,4 +1,4 @@
-from pyspark.sql.types import StructType, StructField, StringType, IntegerType, DateType, TimestampType, FloatType
+from pyspark.sql.types import StructType, StructField, StringType, IntegerType, DateType, DoubleType, FloatType
 
 
 schema = schema = StructType(
@@ -8,9 +8,9 @@ schema = schema = StructType(
         StructField("product_id", StringType(), False),
         StructField("quantity", IntegerType(), True),
         StructField("unit_price", FloatType(), True),
-        StructField("total_amount", FloatType(), True),
+        StructField("total_amount", DoubleType(), True),
         StructField("transaction_date", DateType(), True),
         StructField("transaction_hour", IntegerType(), True)
     ])
-path="/mnt/silver/clean_transactions/data"
-format="delta"
+path = "/mnt/silver/clean_transactions/data"
+format = "delta"
